// src/workspace/__generated__/EditWorkspaceMutation.graphql.ts
export type ProblemType = "BAD_REQUEST" | "CONFLICT" | "FORBIDDEN" | "NOT_FOUND" | "%future added value";
export type UpdateWorkspaceInput = {
  clientMutationId?: string | null;
  description: string;
  maxJobDuration?: number | null;
  metadata?: ResourceMetadataInput | null;
  terraformVersion?: string | null;
  workspacePath: string;
};
export type ResourceMetadataInput = {
  version: string;
};
export type EditWorkspaceMutation$variables = {
  input: UpdateWorkspaceInput;
};
export type EditWorkspaceMutation$data = {
  readonly updateWorkspace: {
    readonly problems: ReadonlyArray<{
      readonly field: ReadonlyArray<string> | null;
      readonly message: string;
      readonly type: ProblemType;
    }>;
    readonly workspace: {
      readonly fullPath: string;
      readonly name: string;
    } | null;
  };
};
export type EditWorkspaceMutation = {
  response: EditWorkspaceMutation$data;
  variables: EditWorkspaceMutation$variables;
};